corrections after review
