+++ packages/client/src/components/ui/Login/LoginPage.module.scss
@@ -30,3 +30,8 @@
 .modalFooter {
   margin-top: 1rem;
 }
+
+.inputForm {
+    display: flex;
+    
+}
