+++ packages/client/src/components/ui/Registration/RegistrationPage.module.scss
