+++ packages/client/src/components/ui/Registration/index.ts
@@ -1,0 +1,1 @@
+export { Registration } from './Registration'
