+++ packages/client/src/styles/index.scss
@@ -1,1 +1,9 @@
 @import 'https://fonts.googleapis.com/css2?family=Golos+Text:wght@400..900&display=swap';
+
+input:-webkit-autofill,
+input:-webkit-autofill:hover,
+input:-webkit-autofill:focus,
+input:-webkit-autofill:active {
+   -webkit-transition-delay: 9999s;
+   transition-delay: 9999s;
+}
